[wip] removing functions and using api_facturae
